requirements.txt
accelerate==0.32.1
deepspeed==0.14.4
diffusers==0.25.1
easydict==1.11
einops==0.8.0
kornia==0.7.1
lpips==0.1.4
matplotlib==3.8.2
mmcv==2.1.0
mmdeploy==1.3.1
mmdet==3.3.0
numpy==2.0.0
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==10.4.0
pycocotools==2.0.8
pytorch_lightning==2.1.3
PyYAML==6.0.1
scipy==1.14.0
setuptools==60.2.0
skimage==0.0
termcolor==2.4.0
terminaltables==3.1.10
timm==0.9.12
torch==2.1.1+cu118
torchmetrics==1.3.0.post0
torchvision==0.16.2
tqdm==4.65.2
transformers==4.37.1
yacs==0.1.8
